+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 - Ref: Remove duplicated attachStackTrace field
 - Fix: Flutter Configurations should be able to mutate the SentryFlutterOptions
+- Enhancement: Add SentryWidgetsBindingObserver, an Integration that captures certain window and device events.
 
 ## 4.0.0-beta.1
 
@@ -18,7 +19,7 @@
 - Enhancement: NavigationObserver to record Breadcrumbs for navigation events #197
 - Fix: Integrations should be closeable
 - Feat: Support split-debug-info for Android #191
-- Fix: The event payload must never serialize null or empty fields
+- Fix: the event payload must never serialize null or empty fields
 - Ref: Make hints optional
 
 ### Breaking changes